requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Templating & Static Files
jinja2==3.1.5
aiofiles==24.1.0

# Database
motor==3.7.0

# AI / LLM
google-generativeai==0.8.4
Pillow==11.1.0

# WhatsApp
twilio==9.4.3

# Email
aiosmtplib==3.0.2

# Scheduling
apscheduler==3.10.4

# Configuration
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Utilities
python-dateutil==2.9.0
pytz==2024.2
httpx==0.28.1
